ArchitectureAgent/requirements.txt
requests
semantic-kernel
azure-search-documents
msal
Pillow
aiohttp
azure-cognitiveservices-vision-computervision
python-dotenv
